moved LoginTrait
