+++ tests/TestCase/src/Controller/BlogPostsControllerTest.php
@@ -16,7 +16,7 @@
 use App\Test\TestCase\AppCakeTestCase;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\IntegrationTestTrait;
-use App\Test\TestCase\LoginTrait;
+use App\Test\TestCase\Traits\LoginTrait;
 
 class BlogPostsControllerTest extends AppCakeTestCase
 {
