+++ tests/TestCase/src/Controller/ManufacturersFrontendControllerTest.php
@@ -15,7 +15,7 @@
  */
 use App\Test\TestCase\AppCakeTestCase;
 use Cake\TestSuite\IntegrationTestTrait;
-use App\Test\TestCase\LoginTrait;
+use App\Test\TestCase\Traits\LoginTrait;
 
 class ManufacturersFrontendControllerTest extends AppCakeTestCase
 {
