+++ tests/TestCase/src/Controller/PagesControllerTest.php
@@ -3,7 +3,7 @@
 use App\Test\TestCase\AppCakeTestCase;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\IntegrationTestTrait;
-use App\Test\TestCase\LoginTrait;
+use App\Test\TestCase\Traits\LoginTrait;
 
 /**
  * PagesControllerTest
